bot.py
import discord
from discord.ext import commands
from dotenv import load_dotenv
import os
import asyncio

load_dotenv()

TOKEN = os.getenv("DISCORD_TOKEN")
GUILD_ID = int(os.getenv("GUILD_ID"))

intents = discord.Intents.default()
bot = commands.Bot(command_prefix="!", intents=intents)

@bot.event
async def on_ready():
    print(f"Logged in as {bot.user} (ID: {bot.user.id})")
    try:
        synced = await bot.tree.sync(guild=discord.Object(id=GUILD_ID))
        print(f"Synced {len(synced)} slash command(s) to dev guild.")
    except Exception as e:
        print(f"Failed to sync: {e}")

async def main():
    async with bot:
        await bot.load_extension("cogs.pingpong")
        await bot.start(TOKEN)

asyncio.run(main())
